requirements.txt
youtube-dl
yt-dlp
discord.py
discord.py[voice]

